Functional support for the :bind and :unbind commands is now complete, though more testing is still required.
This adds support for the following:

    :bind <ACTION> <KEYCODE>

where <ACTION> is one of the available actions supported by vitunes, and
keycode is something recognized by vitunes.  See the upcoming documentation
addition for details on both.


    :unbind *

This un-binds all existing keybindings.  Handy when you want to start from
scratch and define all of your own bindings.


    :unbind action <ACTION>

This unbinds all keys bound to the action specified by <ACTION>.


    :unbind key <KEYCODE>

This unbinds any action currently bound to the keycode specified by <KEYCODE>.

As with any other commands, these can be specified in the vitunes
configuration file.