Adding -c option and :pl command.
With -c it's possible to send a command to a running vitunes instance.
The :pl, or :playlist, command lets you jump to a playlist by name. So,
if you have vitunes running in some terminal, you can execute:

    vitunes -c 'pl Cool_Playlist' -c 'media_play'

to start listening to Cool_Playlist. There is no sugar for the commands
at the moment (you'll have to use the keybindings or command names used
by vitunes internally).

One note about how commands are executed, when send to vitunes via the
-c switch: it will first try to find a keybinding and if that fails just
execute it as a command.