Do not use taglib-config.
Instead, make use of pkg-config unconditionally, since it's pretty much
available everywhere.

Besides, taglib-config does not distinguish from the archive to link
with and a path to search it for, which would all be (wrongly) picked up
by the LDFLAGS variable, because file related options matter to the
linker (the archives must be listed after the object files).

This was prompted by Jasen Price (issue #34). Upcoming commit will
really fix this.