Take advantage of the execute command function.
Now, with error contexts, the execute function can, also, be called when
parsing the configuration file, without having to worry about where to
output its messages.

That is, when parsing the configuration file, the messages are printed to
standard error and, afterwards, they get painted to the command/status
window.